src/data/usecases/AddAccount/dbAddAccount.ts
import {
  AddAccount,
  AddAccountModel,
  AccountModel,
  Encrypter,
  AddAccountRepository,
} from './dbAddAccountProtocols';

export class DbAddAccount implements AddAccount {
  constructor(
    private readonly encrypter: Encrypter,
    private readonly addAccountRepository: AddAccountRepository
  ) {}

  async add(account: AddAccountModel): Promise<AccountModel> {
    const encryptedPassword = await this.encrypter.encrypt(account.password);
    const response = await this.addAccountRepository.add({
      ...account,
      password: encryptedPassword,
    });

    return response;
  }
}
